Criado abstracao de UsuarioDAO e movido metodos relacionados
Criado abstracao de UsuarioDAO e movido metodos relacionados para a respectiva classe
